ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2